Move ACL related code to skydb/access.go
