+++ pkg/server/skydb/access.go
@@ -1,0 +1,128 @@
+// Copyright 2015-present Oursky Ltd.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package skydb
+
+// RecordACLEntry grants access to a record by relation or by user_id
+type RecordACLEntry struct {
+	Relation string   `json:"relation,omitempty"`
+	Role     string   `json:"role,omitempty"`
+	Level    ACLLevel `json:"level"`
+	UserID   string   `json:"user_id,omitempty"`
+	Public   bool     `json:"public,omitempty"`
+}
+
+// ACLLevel represent the operation a user granted on a resource
+type ACLLevel string
+
+// ReadLevel and WriteLevel is self-explanatory
+const (
+	ReadLevel   ACLLevel = "read"
+	WriteLevel           = "write"
+	CreateLevel          = "create"
+)
+
+// NewRecordACLEntryRelation returns an ACE on relation
+func NewRecordACLEntryRelation(relation string, level ACLLevel) RecordACLEntry {
+	return RecordACLEntry{
+		Relation: relation,
+		Level:    level,
+	}
+}
+
+// NewRecordACLEntryDirect returns an ACE for a specific user
+func NewRecordACLEntryDirect(userID string, level ACLLevel) RecordACLEntry {
+	return RecordACLEntry{
+		Relation: "$direct",
+		Level:    level,
+		UserID:   userID,
+	}
+}
+
+// NewRecordACLEntryRole return an ACE on role
+func NewRecordACLEntryRole(role string, level ACLLevel) RecordACLEntry {
+	return RecordACLEntry{
+		Role:  role,
+		Level: level,
+	}
+}
+
+// NewRecordACLEntryPublic return an ACE on public access
+func NewRecordACLEntryPublic(level ACLLevel) RecordACLEntry {
+	return RecordACLEntry{
+		Public: true,
+		Level:  level,
+	}
+}
+
+func (ace *RecordACLEntry) Accessible(userinfo *UserInfo, level ACLLevel) bool {
+	if ace.Public {
+		return ace.AccessibleLevel(level)
+	}
+	if userinfo == nil {
+		return false
+	}
+	if userinfo.ID == ace.UserID {
+		if ace.AccessibleLevel(level) {
+			return true
+		}
+	}
+	for _, role := range userinfo.Roles {
+		if role == ace.Role {
+			if ace.AccessibleLevel(level) {
+				return true
+			}
+		}
+	}
+	return false
+}
+
+func (ace *RecordACLEntry) AccessibleLevel(level ACLLevel) bool {
+	if level == ReadLevel {
+		return true
+	}
+	if level == ace.Level && level == WriteLevel {
+		return true
+	}
+	return false
+}
+
+// RecordACL is a list of ACL entries defining access control for a record
+type RecordACL []RecordACLEntry
+
+// NewRecordACL returns a new RecordACL
+func NewRecordACL(entries []RecordACLEntry) RecordACL {
+	acl := make(RecordACL, len(entries))
+	for i, v := range entries {
+		acl[i] = v
+	}
+	return acl
+}
+
+// Accessible checks whether provided user info has certain access level
+func (acl RecordACL) Accessible(userinfo *UserInfo, level ACLLevel) bool {
+	if len(acl) == 0 {
+		// default behavior of empty ACL
+		return true
+	}
+
+	accessible := false
+	for _, ace := range acl {
+		if ace.Accessible(userinfo, level) {
+			accessible = true
+		}
+	}
+
+	return accessible
+}
